feat: 유저 관련 페이지에 메타데이터 설정
유저 수정 페이지 metadata 설정이 page.tsx에서 동작하지 않는 문제 수정

- page.tsx가 클라이언트 컴포넌트로 취급되어 metadata export 불가
- 해당 세그먼트 layout.tsx(서버 컴포넌트)를 만들고 거기로 metadata 이동
